// src/server/modules/EvolizAPI/interfaces/IEvolizAPIPagination.ts
export default interface IEvolizAPIPagination<EvolizAPIInterface> {
    data: EvolizAPIInterface[];
    links: {
        first: string;
        last: string;
        prev: string;
        next: string;
    };
    meta: {
        current_page: number;
        from: number;
        last_page: number;
        path: string;
        per_page: number;
        to: number;
        total: number;
    };
}
// Generated by Copilot, using the following code as a prompt (from https://evoliz.io/documentation#section/Pagination/Example (v1.24 in https://evoliz.io/changelog)):
// {
//     "data": [
//         // Array of returned resources
//     ],
//         "links": {
//         "first": "https://...?page=1", // link to first page of data
//             "last": "https://...?page=29", // link to last page of data
//                 "prev": null, // link to previous page
//                     "next": "https://...?page=2"  // link to next page
//     },
//     "meta": {
//         "current_page": 1, // Current page number
//             "from": 1,  // List begin on this item number
//                 "last_page": 29, // Last page number
//                     "path": "https://...", // Url to resource without specified page
//                         "per_page": 50, // number of items / page
//                             "to": 50, // List ends on this item number
//                                 "total": 1433 // number of items
//     }
// }